Refactor BlogNavigationController and tour guide
